Интегриране на обработката на изображения в worker.js
