+++ worker.js
@@ -1,4 +1,38 @@
-import { preprocessImage } from './image-utils.js';
+import Jimp from 'jimp';
+
+// --- ПРЕРАБОТКА НА ИЗОБРАЖЕНИЯ ---
+async function resizeToStandard(image, size = 1024) {
+    return image.contain(size, size);
+}
+
+async function autoCropBlackBorders(image) {
+    return image.autocrop({
+        cropSymmetric: true,
+        tolerance: 0.0001,
+        leaveBorder: 0,
+        color: 0x000000FF,
+    });
+}
+
+async function normalizeBrightnessContrast(image) {
+    image.normalize();
+    return image;
+}
+
+async function preprocessImage(file, size = 1024) {
+    try {
+        const buffer = Buffer.from(await file.arrayBuffer());
+        let image = await Jimp.read(buffer);
+        image = await autoCropBlackBorders(image);
+        image = await resizeToStandard(image, size);
+        image = await normalizeBrightnessContrast(image);
+        const outBuffer = await image.getBufferAsync(Jimp.MIME_PNG);
+        return new File([outBuffer], file.name || 'processed.png', { type: 'image/png' });
+    } catch {
+        // Ако обработката се провали, връщаме оригиналния файл
+        return file;
+    }
+}
 
 function validateKv(data) {
   const entries = [];
